+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Remove orphaned resources during dependency analysis.
 - Allow randomly skipping non-body links to test invalid scenarios.
 - Use `query`, `header` & `cookie` in dependency inference.
+- Increased exploration rate for optional parameters in stateful testing.
 
 ## [4.3.11](https://github.com/schemathesis/schemathesis/compare/v4.3.10...v4.3.11) - 2025-10-24